# crates/veronex/src/infrastructure/inbound/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Veronex API",
    "description": "Inference gateway API — LLM routing, job tracking, API key management, and cluster observability.",
    "version": "0.1.0",
    "contact": {
      "name": "Veronex",
      "url": "https://veronex.verobee.com"
    }
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Local dev" }
  ],
  "security": [{ "ApiKeyAuth": [] }],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "SubmitInferenceRequest": {
        "type": "object",
        "required": ["model", "prompt"],
        "properties": {
          "model": { "type": "string", "example": "llama3.2:3b" },
          "prompt": { "type": "string", "example": "Explain hexagonal architecture." },
          "provider_type": { "$ref": "#/components/schemas/ProviderType" },
          "stream": { "type": "boolean", "default": false }
        }
      },
...
      "CreateKeyRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "example": "production-key" },
          "tenant_id": { "type": "string", "nullable": true },
          "rate_limit_rpm": { "type": "integer", "default": 0, "description": "0 = unlimited" },
          "rate_limit_tpm": { "type": "integer", "default": 0, "description": "0 = unlimited" },
          "expires_at": { "type": "string", "format": "date-time", "nullable": true }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string" }
        }
      }
    }
  },
...
  "paths": {
...
    "/v1/inference": {
      "post": {
        "tags": ["inference"],
        "summary": "Submit inference job",
        "operationId": "submitInference",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SubmitInferenceRequest" }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Job accepted",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobResponse" } } }
          },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/v1/inference/{job_id}/stream": {
      "get": {
        "tags": ["inference"],
        "summary": "Stream inference output (SSE)",
        "operationId": "streamInference",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of inference tokens",
            "content": { "text/event-stream": { "schema": { "type": "string" } } }
          },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/v1/inference/{job_id}/status": {
      "get": {
        "tags": ["inference"],
        "summary": "Get job status",
        "operationId": "getJobStatus",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Job status",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobResponse" } } }
          },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/v1/inference/{job_id}": {
      "delete": {
        "tags": ["inference"],
        "summary": "Cancel inference job",
        "operationId": "cancelInference",
        "parameters": [
          { "name": "job_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Job cancelled" },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/v1/chat/completions": {
      "post": {
        "tags": ["openai"],
        "summary": "OpenAI-compatible chat completions",
        "description": "Drop-in replacement for OpenAI's /v1/chat/completions. Supports streaming via SSE.",
        "operationId": "chatCompletions",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/OpenAIChatRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Chat completion response or SSE stream",
            "content": {
              "application/json": { "schema": { "type": "object" } },
              "text/event-stream": { "schema": { "type": "string" } }
            }
          },
          "401": { "description": "Unauthorized" },
          "429": { "description": "Rate limit exceeded" }
        }
      }
    },
    "/v1/keys": {
      "post": {
        "tags": ["keys"],
        "summary": "Create API key",
        "operationId": "createKey",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateKeyRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Key created — raw key shown once",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreatedKeyResponse" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "get": {
        "tags": ["keys"],
        "summary": "List API keys",
        "operationId": "listKeys",
        "responses": {
          "200": {
            "description": "List of API keys",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ApiKeyResponse" } }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/keys/{id}": {
      "delete": {
        "tags": ["keys"],
        "summary": "Delete API key",
        "operationId": "deleteKey",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": { "description": "Key deleted" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Key not found" }
        }
      },
      "patch": {
        "tags": ["keys"],
        "summary": "Toggle API key active state",
        "operationId": "toggleKey",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": { "description": "Key toggled" },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Key not found" }
        }
      }
    },
    "/v1/usage": {
      "get": {
        "tags": ["usage"],
        "summary": "Aggregate usage analytics",
        "operationId": "aggregateUsage",
        "parameters": [
          { "name": "hours", "in": "query", "schema": { "type": "integer", "default": 24 }, "description": "Lookback window in hours" }
        ],
        "responses": {
          "200": {
            "description": "Aggregate usage data",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UsageAggregate" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/usage/{key_id}": {
      "get": {
        "tags": ["usage"],
        "summary": "Per-key usage analytics",
        "operationId": "keyUsage",
        "parameters": [
          { "name": "key_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "hours", "in": "query", "schema": { "type": "integer", "default": 24 } }
        ],
        "responses": {
          "200": {
            "description": "Per-key usage data",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UsageAggregate" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/usage/{key_id}/jobs": {
      "get": {
        "tags": ["usage"],
        "summary": "Jobs list for an API key",
        "operationId": "keyUsageJobs",
        "parameters": [
          { "name": "key_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } }
        ],
        "responses": {
          "200": {
            "description": "Jobs for the key",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/JobResponse" } } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/dashboard/stats": {
      "get": {
        "tags": ["dashboard"],
        "summary": "Cluster-wide statistics",
        "operationId": "getDashboardStats",
        "responses": {
          "200": {
            "description": "Dashboard stats",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/DashboardStats" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/dashboard/jobs": {
      "get": {
        "tags": ["dashboard"],
        "summary": "Paginated job list",
        "operationId": "listDashboardJobs",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0 } },
          { "name": "status", "in": "query", "schema": { "$ref": "#/components/schemas/JobStatus" } },
          { "name": "search", "in": "query", "schema": { "type": "string" }, "description": "Prompt text search" }
        ],
        "responses": {
          "200": {
            "description": "Paginated job list",
            "content": { "application/json": { "schema": { "type": "object", "properties": { "total": { "type": "integer" }, "jobs": { "type": "array", "items": { "$ref": "#/components/schemas/JobResponse" } } } } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/dashboard/jobs/{id}": {
      "get": {
        "tags": ["dashboard"],
        "summary": "Get job detail",
        "operationId": "getDashboardJob",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Job detail",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/JobResponse" } } }
          },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Job not found" }
        }
      }
    },
    "/v1/dashboard/performance": {
      "get": {
        "tags": ["dashboard"],
        "summary": "Performance metrics (latency / throughput)",
        "operationId": "getPerformance",
        "parameters": [
          { "name": "hours", "in": "query", "schema": { "type": "integer", "default": 24 } }
        ],
        "responses": {
          "200": {
            "description": "Performance metrics",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PerformanceMetrics" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/v1/providers": {
      "post": {
        "tags": ["providers"],
        "summary": "Register LLM provider",
        "operationId": "registerProvider",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RegisterProviderRequest" } } }
        },
        "responses": {
          "201": {
            "description": "Provider registered",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProviderResponse" } } }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "get": {
        "tags": ["providers"],
        "summary": "List LLM providers",
        "operationId": "listProviders",
        "responses": {
          "200": {
            "description": "List of providers",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ProviderResponse" } } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
    "/v1/gemini/policies": {
      "get": {
        "tags": ["gemini"],
        "summary": "List Gemini rate-limit policies",
        "operationId": "listGeminiPolicies",
        "responses": {
          "200": {
            "description": "Gemini policies",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/GeminiPolicy" } } } }
          },
          "401": { "description": "Unauthorized" }
        }
      }
    },
...
  }
}